requirements.txt
torch
torchvision
pillow
matplotlib
numpy
pandas
scikit-learn
fastapi
uvicorn
jinja2
python-multipart
